requirements.test.txt
-e .
avisdk
pytest
pytest-cov
pyOpenSSL
codecov
